refactor: main파일에서 실행하던 실행구문 struct BankDesk로 파일분리
